pyproject.toml
[project]
name = "football-tracker"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
